requirements.txt
annotated-types==0.7.0
anyio==4.8.0
av==14.0.1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.5.0
distro==1.9.0
elevenlabs==1.50.4
faster-whisper==1.1.1
filelock==3.17.0
flatbuffers==25.1.21
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
groq==0.15.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
jiter==0.8.2
mpmath==1.3.0
numpy==2.2.2
onnxruntime==1.20.1
openai==1.60.0
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
pyperclip==1.9.0
pyreadline3==3.5.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
SpeechRecognition==3.14.0
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
